Added AlertAddActivity.

+++ app/src/sharedTest/java/com/github/mmodzel3/lostfinder/alert/AlertEndpointTestAbstract.kt
@@ -20,7 +20,7 @@ abstract class AlertEndpointTestAbstract : ServerEndpointTestAbstract() {
         const val ALERT_DESCRIPTION = "description"
         const val ALERT_LONGITUDE = 22.1
         const val ALERT_LATITUDE = 19.2
-        const val ALERT_RANGE = 20.0
+        const val ALERT_RANGE = 20
         const val ALERT_SHOW_NOTIFICATION_AT_START = true
         const val ALERT_SHOW_NOTIFICATION_AT_END = false
         const val DAY_BEFORE_IN_MILLISECONDS = 24*60*60*1000
@@ -53,7 +53,7 @@ abstract class AlertEndpointTestAbstract : ServerEndpointTestAbstract() {
         val yesterday = Date(System.currentTimeMillis() - UserEndpointTestAbstract.DAY_BEFORE_IN_MILLISECONDS)
 
         for (id in 1..4) {
-            alerts.add(Alert(id.toString(), ALERT_TYPE, user, Location(ALERT_LONGITUDE, ALERT_LATITUDE), ALERT_RANGE,
+            alerts.add(Alert(id.toString(), ALERT_TYPE, user, Location(ALERT_LONGITUDE, ALERT_LATITUDE), ALERT_RANGE.toDouble(),
                             ALERT_TITLE, ALERT_DESCRIPTION, ALERT_SHOW_NOTIFICATION_AT_START, ALERT_SHOW_NOTIFICATION_AT_END,
                             yesterday, yesterday, null, yesterday))
         }
